+++ ui/src/hooks/useSessionManagement.ts
@@ -3,7 +3,8 @@ import { useEffect, useCallback, useState } from "react";
 import { useSessionStore } from "@/stores/sessionStore";
 import { useSessionEvents } from "@/hooks/useSessionEvents";
 import { useSettingsStore } from "@/hooks/stores";
-import { usePermissions, Permission } from "@/hooks/usePermissions";
+import { usePermissions } from "@/hooks/usePermissions";
+import { Permission } from "@/types/permissions";
 
 type RpcSendFunction = (method: string, params: Record<string, unknown>, callback: (response: { result?: unknown; error?: { message: string } }) => void) => void;
 
@@ -32,21 +33,19 @@ export function useSessionManagement(sendFn: RpcSendFunction | null) {
     clearSession
   } = useSessionStore();
 
-  const { hasPermission } = usePermissions();
+  const { hasPermission, isLoading: isLoadingPermissions } = usePermissions();
 
   const { requireSessionApproval } = useSettingsStore();
   const { handleSessionEvent } = useSessionEvents(sendFn);
   const [primaryControlRequest, setPrimaryControlRequest] = useState<PrimaryControlRequest | null>(null);
   const [newSessionRequest, setNewSessionRequest] = useState<NewSessionRequest | null>(null);
 
-  // Handle session info from WebRTC answer
   const handleSessionResponse = useCallback((response: SessionResponse) => {
     if (response.sessionId && response.mode) {
       setCurrentSession(response.sessionId, response.mode as "primary" | "observer" | "queued" | "pending");
     }
   }, [setCurrentSession]);
 
-  // Handle approval of primary control request
   const handleApprovePrimaryRequest = useCallback(async (requestId: string) => {
     if (!sendFn) return;
 
@@ -63,7 +62,6 @@ export function useSessionManagement(sendFn: RpcSendFunction | null) {
     });
   }, [sendFn]);
 
-  // Handle denial of primary control request
   const handleDenyPrimaryRequest = useCallback(async (requestId: string) => {
     if (!sendFn) return;
 
@@ -80,7 +78,6 @@ export function useSessionManagement(sendFn: RpcSendFunction | null) {
     });
   }, [sendFn]);
 
-  // Handle approval of new session
   const handleApproveNewSession = useCallback(async (sessionId: string) => {
     if (!sendFn) return;
 
@@ -97,7 +94,6 @@ export function useSessionManagement(sendFn: RpcSendFunction | null) {
     });
   }, [sendFn]);
 
-  // Handle denial of new session
   const handleDenyNewSession = useCallback(async (sessionId: string) => {
     if (!sendFn) return;
 
@@ -114,34 +110,30 @@ export function useSessionManagement(sendFn: RpcSendFunction | null) {
     });
   }, [sendFn]);
 
-  // Handle RPC events
   const handleRpcEvent = useCallback((method: string, params: unknown) => {
-    // Pass session events to the session event handler
     if (method === "sessionsUpdated" ||
         method === "modeChanged" ||
+        method === "connectionModeChanged" ||
         method === "otherSessionConnected") {
       handleSessionEvent(method, params);
     }
 
-    // Handle new session approval request (only if approval is required and user has permission)
-    if (method === "newSessionPending" && requireSessionApproval && hasPermission(Permission.SESSION_APPROVE)) {
-      setNewSessionRequest(params as NewSessionRequest);
+    if (method === "newSessionPending" && requireSessionApproval) {
+      if (isLoadingPermissions || hasPermission(Permission.SESSION_APPROVE)) {
+        setNewSessionRequest(params as NewSessionRequest);
+      }
     }
 
-    // Handle primary control request
     if (method === "primaryControlRequested") {
       setPrimaryControlRequest(params as PrimaryControlRequest);
     }
 
-    // Handle approval/denial responses
     if (method === "primaryControlApproved") {
-      // Clear requesting state in store
       const { setRequestingPrimary } = useSessionStore.getState();
       setRequestingPrimary(false);
     }
 
     if (method === "primaryControlDenied") {
-      // Clear requesting state and show error
       const { setRequestingPrimary, setSessionError } = useSessionStore.getState();
       setRequestingPrimary(false);
       setSessionError("Your primary control request was denied");
@@ -152,9 +144,14 @@ export function useSessionManagement(sendFn: RpcSendFunction | null) {
       const errorParams = params as { message?: string };
       setSessionError(errorParams.message || "Session access was denied by the primary session");
     }
-  }, [handleSessionEvent, hasPermission, requireSessionApproval]);
+  }, [handleSessionEvent, hasPermission, isLoadingPermissions, requireSessionApproval]);
 
-  // Cleanup on unmount
+  useEffect(() => {
+    if (!isLoadingPermissions && newSessionRequest && !hasPermission(Permission.SESSION_APPROVE)) {
+      setNewSessionRequest(null);
+    }
+  }, [isLoadingPermissions, hasPermission, newSessionRequest]);
+
   useEffect(() => {
     return () => {
       clearSession();
